# backend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
      "title": "FastAPI",
      "version": "0.1.0"
    },
    "paths": {
...
      "/addresses/": {
        "get": {
          "tags": [
            "addresses"
          ],
          "summary": "Get Addresses",
          "description": "Get all addresses",
          "operationId": "get_addresses_addresses__get",
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "items": {
                      "$ref": "#/components/schemas/AddressResponse"
                    },
                    "type": "array",
                    "title": "Response Get Addresses Addresses  Get"
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "addresses"
          ],
          "summary": "Create Address",
          "description": "Create a new address",
          "operationId": "create_address_addresses__post",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressCreate"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/AddressResponse"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/addresses/{address_id}": {
        "get": {
          "tags": [
            "addresses"
          ],
          "summary": "Get Address",
          "description": "Get a specific address by ID",
          "operationId": "get_address_addresses__address_id__get",
          "parameters": [
            {
              "name": "address_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Address Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/AddressResponse"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        },
        "put": {
          "tags": [
            "addresses"
          ],
          "summary": "Update Address",
          "description": "Update an existing address",
          "operationId": "update_address_addresses__address_id__put",
          "parameters": [
            {
              "name": "address_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Address Id"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressUpdate"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/AddressResponse"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": [
            "addresses"
          ],
          "summary": "Delete Address",
          "description": "Delete an address",
          "operationId": "delete_address_addresses__address_id__delete",
          "parameters": [
            {
              "name": "address_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Address Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
  
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
        "AddressCreate": {
          "properties": {
            "street": {
              "type": "string",
              "title": "Street"
            },
            "exterior_number": {
              "type": "string",
              "title": "Exterior Number"
            },
            "interior_number": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Interior Number"
            },
            "neighborhood": {
              "type": "string",
              "title": "Neighborhood"
            },
            "postal_code": {
              "type": "string",
              "title": "Postal Code"
            },
            "city": {
              "type": "string",
              "title": "City"
            },
            "state": {
              "type": "string",
              "title": "State"
            },
            "country": {
              "type": "string",
              "title": "Country",
              "default": "México"
            },
            "notes": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Notes"
            },
            "created_at": {
              "type": "string",
              "format": "date-time",
              "title": "Created At"
            },
            "updated_at": {
              "type": "string",
              "format": "date-time",
              "title": "Updated At"
            }
          },
          "type": "object",
          "required": [
            "street",
            "exterior_number",
            "neighborhood",
            "postal_code",
            "city",
            "state"
          ],
          "title": "AddressCreate"
        },
        "AddressResponse": {
          "properties": {
            "street": {
              "type": "string",
              "title": "Street"
            },
            "exterior_number": {
              "type": "string",
              "title": "Exterior Number"
            },
            "interior_number": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Interior Number"
            },
            "neighborhood": {
              "type": "string",
              "title": "Neighborhood"
            },
            "postal_code": {
              "type": "string",
              "title": "Postal Code"
            },
            "city": {
              "type": "string",
              "title": "City"
            },
            "state": {
              "type": "string",
              "title": "State"
            },
            "country": {
              "type": "string",
              "title": "Country",
              "default": "México"
            },
            "notes": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Notes"
            },
            "created_at": {
              "type": "string",
              "format": "date-time",
              "title": "Created At"
            },
            "updated_at": {
              "type": "string",
              "format": "date-time",
              "title": "Updated At"
            },
            "id": {
              "type": "integer",
              "title": "Id"
            }
          },
          "type": "object",
          "required": [
            "street",
            "exterior_number",
            "neighborhood",
            "postal_code",
            "city",
            "state",
            "id"
          ],
          "title": "AddressResponse"
        },
        "AddressUpdate": {
          "properties": {
            "street": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Street"
            },
            "exterior_number": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Exterior Number"
            },
            "interior_number": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Interior Number"
            },
            "neighborhood": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Neighborhood"
            },
            "postal_code": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Postal Code"
            },
            "city": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "City"
            },
            "state": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            },
            "country": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Country"
            },
            "notes": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Notes"
            }
          },
          "type": "object",
          "title": "AddressUpdate"
        },
...
        "HTTPValidationError": {
          "properties": {
            "detail": {
              "items": {
                "$ref": "#/components/schemas/ValidationError"
              },
              "type": "array",
              "title": "Detail"
            }
          },
          "type": "object",
          "title": "HTTPValidationError"
        },
...
        "ValidationError": {
          "properties": {
            "loc": {
              "items": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "integer"
                  }
                ]
              },
              "type": "array",
              "title": "Location"
            },
            "msg": {
              "type": "string",
              "title": "Message"
            },
            "type": {
              "type": "string",
              "title": "Error Type"
            }
          },
          "type": "object",
          "required": [
            "loc",
            "msg",
            "type"
          ],
          "title": "ValidationError"
        }
      }
    }
  }